AuditHandlerRegistry as bean, improved log messages
