+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/auditview/control/NoAuditHandlerException.java
@@ -1,0 +1,6 @@
+package ch.puzzle.itc.mobiliar.business.auditview.control;
+
+public class NoAuditHandlerException extends Exception {
+
+    private static final long serialVersionUID = 1L;
+}
